% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-round-rect.R
\name{draw_key_round_rect}
\alias{draw_key_round_rect}
\title{图例}
\usage{
draw_key_round_rect(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A grob representing the key.
}
\description{
图例
}
