% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{element_round_rect}
\alias{element_round_rect}
\title{: Round rectangle element}
\usage{
element_round_rect(
  fill = NULL,
  colour = NULL,
  linewidth = NULL,
  linetype = NULL,
  color = NULL,
  inherit.blank = FALSE,
  radius = 2
)
}
\arguments{
\item{fill}{Fill colour.}

\item{colour, color}{Line/border colour. Color is an alias for colour.}

\item{linewidth}{Line/border size in mm.}

\item{linetype}{Line type.}

\item{inherit.blank}{Should this element inherit the existence of an
\code{element_blank} among its parents? If \code{TRUE} the existence of
a blank element among its parents will cause this element to be blank as
well. If \code{FALSE} any blank parent element will be ignored when
calculating final element state.}

\item{radius}{The radius of the corners of the rectangle.}
}
\value{
An S3 object of class \code{element}, \code{rel}, or \code{margin}.
}
\description{
In conjunction with the \link{theme} system, the \code{element_} functions
specify the display of how non-data components of the plot are drawn.
\itemize{
\item \code{element_blank()}: draws nothing, and assigns no space.
\item \code{element_rect()}: borders and backgrounds.
\item \code{element_line()}: lines.
\item \code{element_text()}: text.
}

\code{rel()} is used to specify sizes relative to the parent,
\code{margin()} is used to specify the margins of elements.
}
\examples{
plot <- ggplot(mpg, aes(displ, hwy)) + geom_point()


plot + theme(
  panel.background = element_round_rect(fill = "white"),
  plot.margin = margin(2, 2, 2, 2, "cm"),
  plot.background = element_round_rect(
    radius = grid::unit(1, "mm"),
    fill = "grey90",
    colour = "black",
    linewidth = 1,
    linetype = 1
  )
)
}
