% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-round-corssbar.R
\name{geom_round_crossbar}
\alias{geom_round_crossbar}
\title{round crossbar}
\usage{
geom_round_crossbar(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  fatten = 2.5,
  radius = 2,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()} or \code{aes_()}.}

\item{data}{A layer specific dataset - only needed if you want to override the
plot defaults.}

\item{stat}{The statistical transformation to use on the data for this layer,
as a string.}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{layer()}.}

\item{fatten}{A multiplicative factor to fatten the line ends.}

\item{radius}{The radius of the rounded corners.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning.}

\item{orientation}{The orientation of the layer. The default (NA) automatically
determines the orientation from the aesthetic mappings.}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{If FALSE, overrides the default aesthetics.}
}
\description{
round crossbar
}
\details{
\code{geom_round_crossbar()} is a cross between \code{geom_round_tile()} and
\code{geom_round_bar()}. It is a bar that is rounded at the ends, but it is not
filled in. It is useful for showing the range of values in a dataset.
}
