% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-round-half-boxplot.R
\name{geom_round_half_boxplot}
\alias{geom_round_half_boxplot}
\title{A half boxplot}
\usage{
geom_round_half_boxplot(
  mapping = NULL,
  data = NULL,
  stat = "boxplot",
  position = "dodge2",
  ...,
  side = "l",
  center = FALSE,
  nudge = 0,
  radius = grid::unit(2, "pt"),
  outlier.colour = NULL,
  outlier.color = NULL,
  outlier.fill = NULL,
  outlier.shape = 19,
  outlier.size = 1.5,
  outlier.stroke = 0.5,
  outlier.alpha = NULL,
  notch = FALSE,
  notchwidth = 0.5,
  varwidth = FALSE,
  errorbar.draw = TRUE,
  errorbar.length = 0.5,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()} or \code{aes_()}.}

\item{data}{A layer specific dataset - only needed if you want to override the}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{layer()}. These are often aesthetics,}

\item{side}{The side of the half-geom, "l" for left and "r" for right, defaults to "l".}

\item{center}{Boolean whether to center the half-boxplot instead of aligning it to its respective side.}

\item{nudge}{Add space between the boxplot and the middle of the space allotted to a given factor on the x-axis.}

\item{radius}{The radius of the rounded corners of the half-boxplot.}

\item{outlier.colour}{The color of the outliers.}

\item{outlier.fill}{The fill color of the outliers.}

\item{outlier.shape}{The shape of the outliers.}

\item{outlier.size}{The size of the outliers.}

\item{outlier.stroke}{The stroke of the outliers.}

\item{outlier.alpha}{The alpha of the outliers.}

\item{notch}{Boolean whether to draw a notch in the half-boxplot.}

\item{notchwidth}{The width of the notch.}

\item{varwidth}{Boolean whether to adjust the width of the half-boxplot based on the number of observations.}

\item{errorbar.draw}{Draw horizontal whiskers at the top and bottom (the IQR). Defaults to \code{TRUE}.}

\item{errorbar.length}{Length of the horizontal whiskers (errorbar). Defaults to half the width of the half-boxplot}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a warning.}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics.}
}
\value{
A ggplot2 layer.
}
\description{
A half boxplot
}
\examples{
ggplot(iris, aes(x = Species, y = Petal.Width, fill = Species)) +
  geom_round_half_boxplot()

ggplot(iris, aes(x = Species, y = Petal.Width, fill = Species)) +
  geom_round_half_boxplot(side = "r")

ggplot(iris, aes(x = Species, y = Petal.Width, fill = Species)) +
  geom_round_half_boxplot(center = TRUE)
}
